Adding in some final touches
1. Changing Stats UI to be percentage based
2. Adding in Win/Lose screen animations
3. Updating data folder structure
4. Created Light Ninja AI

+++ Assets/Scripts/Interface/IHealth.cs
@@ -8,7 +8,7 @@ namespace AD.Interfaces
 {
     public interface IHealth
     {
-        public Action<int> OnAmountChange { get; set; }
+        public Action<float> OnAmountChange { get; set; }
         public Action OnAmountEqualsZero { get; set; }
         public float Amount { get; set; }
         public void AddAmount(float amount);
